Change variable "y" in Resource.fmu to Integer, (#42)
explicitly set variability for variable "y", and add missing Initial Unknown to model structure

see #36

+++ Resource/model.c
@@ -69,7 +69,7 @@ void calculateValues(ModelInstance *comp) {
     free(path);
 }
 
-Status getFloat64(ModelInstance* comp, ValueReference vr, double *value, size_t *index) {
+Status getInt32(ModelInstance* comp, ValueReference vr, int *value, size_t *index) {
     switch (vr) {
         case vr_y:
             value[(*index)++] = M(y);
